Handle coordinates

+++ manager-database/src/main/java/pt/unl/fct/miei/usmanagement/manager/database/hosts/Coordinates.java
@@ -38,7 +38,7 @@ import java.io.Serializable;
 public class Coordinates implements Serializable {
 
 	private final String label;
-	private final float latitude;
-	private final float longitude;
+	private final double latitude;
+	private final double longitude;
 
 }
